TESTS: move polars clinical kernel tests into test_clinical_kernel
polars calls pass ordinal_categories explicitly (there is no ordered-dtype
auto-detection for polars), so the tests move verbatim rather than sharing
a driver with the pandas variants. Collected test count (71) and coverage
are unchanged.
